requirements.txt
asgiref==3.7.2
Django==4.2.6
django-modeltranslation==0.18.11
path==16.7.1
Pillow==10.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.10.0
